services/platform/backgroud-removal-service/pyproject.toml
[project]
name = "background-removal-service"
version = "1.0.0"
description = "Production-ready background removal service with text-based object detection"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",
    "pillow>=10.4.0",
    "numpy>=2.1.0",
    "opencv-python>=4.10.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.1",
    "structlog>=24.4.0",
    # PyTorch ROCm - install manually, see README
    # torch>=2.4.0
    # torchvision>=0.19.0
    "transformers>=4.45.0",
    "scipy>=1.14.0",
    "timm>=1.0.0",
    "safetensors>=0.4.5",
    "huggingface-hub>=0.25.0",
    # SAM2 and GroundingDINO will be installed from git
    # sam2 @ git+https://github.com/facebookresearch/segment-anything-2.git
    # groundingdino @ git+https://github.com/IDEA-Research/GroundingDINO.git
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "black>=24.8.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["setuptools>=75.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
